# Add option to package classes directory to a jar on the compile classpath