--- a/subprojects/docs/src/docs/userguide/java_library_plugin.adoc
+++ b/subprojects/docs/src/docs/userguide/java_library_plugin.adoc
@@ -249,3 +249,11 @@ include::sample[dir="java-library/with-groovy/kotlin",files="a/build.gradle.kts[
 === Increased memory usage for consumers
 
 When a project uses the Java Library plugin, consumers will use the output classes directory of this project directly on their compile classpath, instead of the jar file if the project uses the Java plugin. An indirect consequence is that up-to-date checking will require more memory, because Gradle will snapshot individual class files instead of a single jar. This may lead to increased memory consumption for large projects.
+
+[[sec:java_library_known_issues_windows_performance]]
+=== Significant build performance drop on Windows for huge multi-projects
+
+Another side effect of the snapshotting of individual class files, only affecting Windows systems, is that the performance can significantly drop when processing a very large amount of class files on the compile classpath.
+This only concerns very large multi-projects where a lot of classes are present on the classpath by using many `api` or (deprecated) `compile` dependencies.
+To mitigate this, you can set the `org.gradle.java.compile-classpath-packaging` system property to `true` to change the behavior of the Java Library plugin to use jars instead of class folders for everything on the compile classpath.
+Note, since this has other performance impacts and potentially side effects, by triggering all jar tasks at compile time, it is only recommended to activate this if you suffer from the described performance issue on Windows.