--- a/subprojects/plugins/src/integTest/groovy/org/gradle/java/fixtures/JavaTestFixturesIntegrationTest.groovy
+++ b/subprojects/plugins/src/integTest/groovy/org/gradle/java/fixtures/JavaTestFixturesIntegrationTest.groovy
@@ -21,4 +21,10 @@ class JavaTestFixturesIntegrationTest extends AbstractJavaTestFixturesIntegratio
     String getPluginName() {
         'java'
     }
+
+    @Override
+    List getSkippedJars(boolean compileClasspathPackaging) {
+        compileClasspathPackaging ? [] : [':testFixturesJar']
+    }
+
 }